xtreemfs_java_generator: getInterfaceVersion, getOperationNumber
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@133 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ servers/src/org/xtreemfs/interfaces/Constants.java
@@ -3,11 +3,11 @@ package org.xtreemfs.interfaces;
 
 public interface Constants
 {
-    public static final long STRIPING_POLICY_DEFAULT = 0;
-    public static final long STRIPING_POLICY_RAID0 = 1;
-    public static final long ACCESS_CONTROL_POLICY_NULL = 1;
-    public static final long ACCESS_CONTROL_POLICY_POSIX = 2;
-    public static final long ACCESS_CONTROL_POLICY_VOLUME = 3;
-    public static final long OSD_SELECTION_POLICY_SIMPLE = 1;
-    public static final long PARTITIONING_POLICY_ONE_SLICE_PER_VOLUME = 1;
+    public static final int STRIPING_POLICY_DEFAULT = 0;
+    public static final int STRIPING_POLICY_RAID0 = 1;
+    public static final int ACCESS_CONTROL_POLICY_NULL = 1;
+    public static final int ACCESS_CONTROL_POLICY_POSIX = 2;
+    public static final int ACCESS_CONTROL_POLICY_VOLUME = 3;
+    public static final int OSD_SELECTION_POLICY_SIMPLE = 1;
+    public static final int PARTITIONING_POLICY_ONE_SLICE_PER_VOLUME = 1;
 };
